[IMP] bind this correctly and rename object model, app
